requirements.txt
Flask==1.1.2
gunicorn==20.0.4
pylint==2.6.0
pylint-quotes==0.2.1
PyNaCl==1.4.0
wg-conf==0.5.8
